[CBRD-22024] performance statistics collecting interface (#1065)
http://jira.cubrid.org/browse/CBRD-22024

This is a refactoring/enhancement task resulted from trying to clean up the performance statistics of thread worker pool and daemons. This results in an interface to help collecting statistics optimally (performance and code management wise).

The requirements:

  1. quick increment & time operations.
  2. common types for statistics values and clocking.
  3. a similar interface to collect non-atomic and atomic statistics.
  4. consistent management of multiple statistics in a module regarding statistic types, accepted operations on the type and names of collected statistics.

The solution:

  The first challenge was to find a way to manage **non-atomic and atomic statistics**. First solution was to duplicate entire implementation. All structures and operations were duplicated. I changed that by using a boolean template IsAtomic. All types are defined as:
  ```c
  template<bool IsAtomic>
  class generic_customtype
  {
    ...
  };
  using customtype = generic_customtype<false>;  // non-atomic type
  using atomic_customtype = generic_customtype<true>; // atomic type
  ```

Second challenge was how to **consistently define multiple statistics** for a module/class without messing up statistics types, operations and names. My initial statistics collecting implementation on worker pool and daemons repeated a lot of code for member declarations to store statistics values, reading the values and collecting in performance monitor. To avoid part of that, I added stat_definition and statset_definition classes, inspired by the model we use in performance monitor to collect global statistics.
With the help of statset_definition, all statistics definitions are kept in one place. For a new statistic, just update statset_definition. Then you can call the right operations on new statistic in module/class implementation.
I could not remove duplicated performance_monitor code due to client not having access to server modules. Therefore names & value count are also defined in performance_monitor.c.
The statset_definition class verifies operations are used correctly (according to defined statistics type), manages the value storage (creating sets of the right size), keeps names of statistics (in a correct order).
Initialize statset_definition using initialization array:
  ```c
  static const statset_definition =
    {
      stat_definition (...),
      ....
      stat_definition (...)
    };
```

Third consideration was **optimized statistic operations**. There are two type of statistic value - counters and timers, and two types of collecting methods - atomic and non-atomic. There are three type of statistic definitions - counter, time and counter/timer.
The operations are -

  1. Counter: ```increment (counter default 1);```
  2. Timer: ```time (duration d);``` and ```time (void); // incremented based in internal time point that is reset on every call```
  3. Counter/timer: ``time_and_increment`` which is equivalent to calling time + increment.

Note: there is an overhead compared the dummy way of defining all statistics as members. The offset to statistic value storage is looked up in statset_definition - one extra memory read. I hope we can afford this price for consistence.

Entire list of changes:

  1. moved get_clockfreq from perf.c to tsc_timer.c (the only place it was used).
  2. created perf.cpp, perf.hpp and perf_def.hpp.
  3. perf_def.hpp is an interface to include in header files of modules/classes that require statistics collection. it contains definitions of clocking types, statistics and statistics set types. Required to define statset references, and value getters.
  4. perf.hpp is an interface to include in implementation files of modules/class that require statistics collection. it contains stat_definition/statset_definition classes and the implementation of template/inline functionality. Required to define and create statistic sets and to call statistic operations.
  5. perf.cpp is the implementation of non-template/non-inline functionality.
  6. implemented new system in thread_waiter, thread_looper, thread_daemon and thread_worker_pool.
  7. removed template and moved cubthread::waiter::wait_for/wait_until function to implementation file.
  8. removed duration representation template and moved cubthread::looper constructors to implementation file. all constructors now call default constructor. fixed function binding issues (pass reference of *this, previously copy-constructed on bind; fixing looper copy constructor assigning binder directly).
  9. fixed compiler warnings.
  10. fixed thread manager unit test.

+++ src/query/query_manager.c
@@ -326,7 +326,7 @@ qmgr_allocate_query_entry (THREAD_ENTRY * thread_p, QMGR_TRAN_ENTRY * tran_entry
 	   * This may help us to quickly locate an available id.
 	   */
 	  assert (hint_query_id != 0);
-	  tran_entry_p->query_id_generator = hint_query_id;
+	  tran_entry_p->query_id_generator = (int) hint_query_id;
 	}
     }
   assert (usable == true);
